services/image_gen/ai_image_gen/requirements.txt
fastapi
uvicorn[standard]
kafka-python
SQLAlchemy
mysql-connector-python
pydantic-settings
diffusers==0.32.2
transformers
accelerate
torch==2.4.0
requests
tqdm
python-dotenv
fastapi
pydantic
Pillow
PyYAML
uvicorn
sentencepiece
peft
gunicorn
einops==0.8.0
timm